+++ src/webcouturier/dropdownmenu/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from AccessControl import allow_module
 from zope.i18nmessageid import MessageFactory
 
 msg_fact = MessageFactory('webcouturier.dropdownmenu')
-allow_module('webcouturier.dropdownmenu.utils')